Recovery developer test and add QA test
